requirements.txt
fastapi[all]
openai
pandas
sqlalchemy
pydantic
python-dotenv
psycopg2